Combine schemas into schema.py
